pyproject.toml
[tool.poetry]
name = "medpoisk-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
faker = "^19.12.0"
faker-commerce = "^1.0.3"
httpx = "^0.25.0"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.6"
pydantic-settings = "^2.1.0"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
pre-commit = "^3.6.0"
sqlalchemy-repr = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pytest = "^7.4.2"
mypy = "^1.8.0"
ruff = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = ["F","E","W","I"]

ignore  = ["F403"]
